+++ nexus/test-utils/src/resource_helpers.rs
@@ -2,6 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+use crate::ControlPlaneTestContext;
+
 use super::http_testing::dropshot_compat::objects_post;
 use super::http_testing::AuthnMode;
 use super::http_testing::NexusRequest;
@@ -13,8 +15,12 @@ use omicron_common::api::external::ByteCount;
 use omicron_common::api::external::Disk;
 use omicron_common::api::external::IdentityMetadataCreateParams;
 use omicron_common::api::external::VpcRouter;
+use omicron_nexus::crucible_agent_client::types::State as RegionState;
 use omicron_nexus::external_api::params;
 use omicron_nexus::external_api::views::{Organization, Project, Vpc};
+use omicron_sled_agent::sim::SledAgent;
+use std::sync::Arc;
+use uuid::Uuid;
 
 pub async fn objects_list_page_authz<ItemType>(
     client: &ClientTestContext,
@@ -199,4 +205,39 @@ pub async fn project_get(
         .expect("failed to get project")
         .parsed_body()
         .expect("failed to parse Project")
+}
+
+pub struct DiskTest {
+    pub sled_agent: Arc<SledAgent>,
+    pub zpool_id: Uuid,
+    pub zpool_size: ByteCount,
+    pub dataset_ids: Vec<Uuid>,
+}
+
+impl DiskTest {
+    // Creates fake physical storage, an organization, and a project.
+    pub async fn new(cptestctx: &ControlPlaneTestContext) -> Self {
+        let sled_agent = cptestctx.sled_agent.sled_agent.clone();
+
+        // Create a Zpool.
+        let zpool_id = Uuid::new_v4();
+        let zpool_size = ByteCount::from_gibibytes_u32(10);
+        sled_agent.create_zpool(zpool_id, zpool_size.to_bytes()).await;
+
+        // Create multiple Datasets within that Zpool.
+        let dataset_count = 3;
+        let dataset_ids: Vec<_> =
+            (0..dataset_count).map(|_| Uuid::new_v4()).collect();
+        for id in &dataset_ids {
+            sled_agent.create_crucible_dataset(zpool_id, *id).await;
+
+            // By default, regions are created immediately.
+            let crucible = sled_agent.get_crucible_dataset(zpool_id, *id).await;
+            crucible
+                .set_create_callback(Box::new(|_| RegionState::Created))
+                .await;
+        }
+
+        Self { sled_agent, zpool_id, zpool_size, dataset_ids }
+    }
 }
